% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCounts.R
\name{calcCounts}
\alias{calcCounts}
\title{Creating the count matrix using the calculated bins}
\usage{
calcCounts(pD, bins)
}
\description{
This function will return a matrix object where ecah column is a sample, and each row is the count in a bin
}
\examples{
calcCounts() 
}
\keyword{calcCounts}
