% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DnMD-package.R
\docType{package}
\name{DnMD-package}
\alias{DnMD-package}
\alias{DnMD}
\title{The DnMD package for Detection of Denovo Deletions from Targeted Sequencing Data Using Minimum-Distance}
\description{
The DnMD package for Detection of Denovo Deletions from Targeted Sequencing Data Using Minimum-Distance
}
