% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{This function calls CBS on a matrix of minimum distances to calculate copy number segments}
\usage{
filter(cbs)
}
\arguments{
\item{md}{The minimum distance matrix produced by calcMD}

\item{bins}{The set of bins determined by calcBins}
}
\description{
This function will return a single GRanges object containing the copy number segments of all families in input
}
\keyword{filter}
