% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCNP.R
\name{findCNP}
\alias{findCNP}
\title{Reports probable CNP regions}
\usage{
findCNP(mCounts, bins, pData, plot = F)
}
\arguments{
\item{mCounts}{A matrix of normalized coverage output by normalizedCounts().}

\item{bins}{The set of bins determined by calcBins().}

\item{pData}{A table in the format of the output of pData().}
}
\description{
This function will report likely CNPs and most likely copy number states.
}
\keyword{=}
\keyword{bins,}
\keyword{calcMD}
\keyword{calcMD(mCounts,}
\keyword{md}
\keyword{pD)}
