% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DnMD-package.R
\docType{package}
\name{MDTS-package}
\alias{MDTS-package}
\alias{MDTS}
\title{The MDTS package for Detection of Denovo Deletions from Targeted Sequencing Data Using Minimum-Distance}
\description{
The MDTS package for Detection of Denovo Deletions from Targeted Sequencing Data Using Minimum-Distance
}
