% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCounts.R
\name{normalizeCounts}
\alias{normalizeCounts}
\title{Normalizes the raw coverage counts to M scores}
\usage{
normalizeCounts(countMat, bins)
}
\arguments{
\item{countMat}{A matrix of raw coverage output by calcCounts() where each column is a sample, and each row a bin}

\item{bins}{The set of bins determined by calcBins}
}
\description{
This function will return a matrix object where ecah column is a family, and each row is the M score in a bin
}
\keyword{normalizeCounts}
