% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPT.R
\name{NPT_plot}
\alias{NPT_plot}
\title{Plot results from nested principal component analysis}
\usage{
NPT_plot(pca_obj, group = NULL)
}
\arguments{
\item{pca_obj}{An rda object returned by the second-level PCA in the NPT function.}

\item{group}{A vector or factor specifying the grouping of samples.}
}
\value{
A ggplot object representing a biplot of species scores and sample points.
}
\description{
This function creates a visualization of the results from the nested principal
component analysis (NPT function) for ecological niche periodicity.
}
\details{
The function creates a biplot that includes:
\enumerate{
\item Sample points colored by group
\item Species scores represented as arrows
\item Species labels positioned using ggrepel to avoid overlapping
}

The plot also includes dashed lines at x=0 and y=0, and displays the percentage
of variance explained by each principal component on the axes.
}
\examples{
data(PFF)
PFF[,3:20] <- log(PFF[,3:20])
PFF <- na.omit(PFF)
traits_dimension <- list(
  grow = c("SLA","Leaf_area","LDMC","SRL","Leaf_Nmass","Leaf_Pmass","Root_Nmass"),
  survive = c("Height","Leaf_Cmass","Root_Cmass","Leaf_CN","Leaf_NP","Leaf_CP","Root_CN"),
  reproductive = c("SeedMass","FltDate","FltDur")
)
npt_result <- NPT(data = PFF, dimension = traits_dimension)
NPT_plot(npt_result$result, PFF$family)

}
