% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentMD.R
\name{segmentMD}
\alias{segmentMD}
\title{Obtain segmented minimum distances}
\usage{
segmentMD(md, bins)
}
\arguments{
\item{md}{The minimum distance matrix produced by calcMD}

\item{bins}{The set of bins determined by calcBins}
}
\description{
This function will return a GRanges object that contains segmented minimum distances
}
\keyword{segmentMD}
