% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TN.R
\name{TN_plot}
\alias{TN_plot}
\title{Plot Trait Network Graph}
\usage{
TN_plot(graph, style = 1)
}
\arguments{
\item{graph}{An igraph object representing the trait network.}

\item{style}{A numeric value that determines the plotting style (default is 1).}
}
\description{
This function visualizes the trait network graph generated by the \code{TN} function.
}
\details{
The function uses the \code{cluster_fast_greedy} algorithm to identify communities
in the graph and assigns community membership to vertices. It offers two
plotting styles:
\itemize{
\item Style 1: Plots the community structure.
\item Style 2: Plots the graph in a circular layout with vertex colors representing communities.
}
}
\examples{
data(WH)
WH <- WH[,4:23]
Tn_result <- TN(traits_matrix = WH, rThres = 0.2, pThres = 0.05)
TN_plot(Tn_result, style = 1)
TN_plot(Tn_result, style = 2)

}
