% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBins.R
\name{calcBins}
\alias{calcBins}
\title{Sample driven calculation of bins for counting}
\usage{
calcBins(pD, n, rl, med, min, genome, map_file, seed = 1337)
}
\arguments{
\item{pD}{A table in the format of the output of pData()}

\item{n}{The number of subsamples to use}

\item{med}{The median number of reads across sub-samples to reach before creating a new bin}

\item{min}{The miminum number of coverage across all sub-samples required to create the proto-region}

\item{genome}{The BSGenome object that assists in calculations of the GC content of the bins}

\item{map_file}{A path to the bigwig file of 100mer mappability of the corresponding genome}

\item{seed}{Sets the seed so results are reproducible. Defaults to 1337}
}
\description{
This function will randomly select a sample of bam files to calculate a range of bins to calculate read-depth on
}
\examples{
calcBins() 
}
\keyword{calcBins}
