% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBins.R
\name{calcBins}
\alias{calcBins}
\title{Sample driven calculation of bins for counting}
\usage{
calcBins(pD, n, rl, med, min, genome, map_file, seed = 1337)
}
\description{
This function will randomly select a sample of bam files to calculate a range of bins to calculate read-depth on
}
\examples{
calcBins() 
}
\keyword{calcBins}
