% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCounts.R
\name{calcCounts}
\alias{calcCounts}
\title{Creating the count matrix using the calculated bins}
\usage{
calcCounts(pD, bins)
}
\arguments{
\item{pD}{A table in the format of the output of pData()}

\item{bins}{The set of bins determined by calcBins}
}
\description{
This function will return a matrix object where ecah column is a sample, and each row is the count in a bin
}
\keyword{calcCounts}
