% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCounts.R
\name{calcCounts}
\alias{calcCounts}
\title{Creating the raw count matrix}
\usage{
calcCounts(pData, bins, rl, mc.cores = 1)
}
\arguments{
\item{pData}{A table in the format of the output of pData().}

\item{bins}{The set of bins determined by calcBins().}

\item{rl}{The read length of the experiment.}

\item{mc.cores}{The number of cores to use for multi-threaded analysis. Defaults to 1.}
}
\description{
This function will return a matrix of read counts where ecah column is a sample, and each row is a bin.
}
\keyword{calcCounts}
