% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMD.R
\name{calcMD}
\alias{calcMD}
\title{Calculating a minimum-distance matrix from raw counts, bins, and pData}
\usage{
calcMD(counts, bins, pData)
}
\arguments{
\item{bins}{The set of bins determined by calcBins}

\item{pData}{A table in the format of the output of pData()}
}
\description{
This function will return a matrix object where ecah column is a sample, and each row is the count in a bin
}
\keyword{calcMD}
