% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMD.R
\name{calcMD}
\alias{calcMD}
\title{Calculating the Minimum Distance matrix}
\usage{
calcMD(mCounts, bins, pData)
}
\arguments{
\item{mCounts}{A matrix of normalized coverage output by normalizedCounts().}

\item{bins}{The set of bins determined by calcBins().}

\item{pData}{A table in the format of the output of pData().}
}
\value{
A \code{data.frame} of minimum distances. Each column is a trio,
while each row is an entry in \code{bins}
}
\description{
This function will return a matrix of minimum distances where ecah column is 
a family, and each row is a bin.
}
\examples{
setwd(system.file('extdata', package='MDTS'))
load('bins.RData')
load('counts.RData')
load('pD.RData')
mCounts = normalizeCounts(counts, bins)
md = calcMD(mCounts, bins, pD)
}
\keyword{calcMD}
