% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denovoDeletions.R
\name{denovoDeletions}
\alias{denovoDeletions}
\title{Denovo Deletion Calling}
\usage{
denovoDeletions(cbs, mCounts, bins)
}
\arguments{
\item{cbs}{The GRanges output from segmentMD() that contains the copy number segments of the families.}

\item{mCounts}{The normalized counts matrix output by normalizeCounts().}

\item{bins}{The set of bins determined by calcBins().}
}
\description{
This function will return a single GRanges object containing all denovo deletions 
		that passed filtering from a Circular Binary Segmentation object with supplementary information.
}
\examples{
setwd(system.file('extdata', package='MDTS'))
load('bins.RData')
load('counts.RData')
load('pD.RData')
mCounts = normalizeCounts(counts, bins)
md = calcMD(mCounts, bins, pD)
cbs = segmentMD(md, bins)
denovo = denovoDeletions(cbs, mCounts, bins)
}
\keyword{denovoDeletions}
