% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCounts.R
\name{normalizeCounts}
\alias{normalizeCounts}
\title{Calculating the normalized M scores}
\usage{
normalizeCounts(counts, bins, mc.cores = 1)
}
\arguments{
\item{counts}{A matrix of raw coverage output by calcCounts().}

\item{bins}{The set of bins determined by calcBins().}
}
\description{
This function will return a matrix of normalized M scores where ecah column is a sample, and each row is a bin.
}
\keyword{normalizeCounts}
