% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCounts.R
\name{normalizeCounts}
\alias{normalizeCounts}
\title{Calculating the normalized M scores}
\usage{
normalizeCounts(counts, bins, mc.cores = 1)
}
\arguments{
\item{counts}{A matrix of raw coverage output by calcCounts().}

\item{bins}{The set of bins determined by calcBins().}

\item{mc.cores}{The number of cores to use for multi-threaded analysis. 
Defaults to 1.}
}
\value{
A \code{data.frame} of normalized counts. Each column is a sample,
and each row is a entry of \code{bins}.
}
\description{
This function will return a matrix of normalized M scores where ecah column 
is a sample, and each row is a bin.
}
\examples{
setwd(system.file('extdata', package='MDTS'))
load('bins.RData')
load('counts.RData')
load('pD.RData')
mCounts = normalizeCounts(counts, bins)
}
\keyword{normalizeCounts}
