% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pData.R
\name{pData}
\alias{pData}
\title{Constructor for pData}
\usage{
pData(path)
}
\arguments{
\item{path}{The path to tab-delimited file storing the phenotype information.}
}
\value{
Returns a \code{data.frame} of required sample information for 
running MDTS.
}
\description{
This function allows constructor of phenotype information necessary for 
downstream analysis. See format of required fields. Function will also 
rearrange the rows such that trios are grouped together - with proband first,
mother second, and father third.
}
\examples{
system.file('extdata', package='MDTS')
pD = pData('https://raw.githubusercontent.com/JMF47/MDTSData/master/data/pD.ped')
}
