% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentMD.R
\name{segmentMD}
\alias{segmentMD}
\title{Circular Binary Segmentation on Minimum Distances}
\usage{
segmentMD(md, bins)
}
\arguments{
\item{md}{The minimum distance matrix produced by calcMD.}

\item{bins}{The set of bins determined by calcBins.}
}
\description{
This function will return a GRanges object containing the copy number 
segments of all families in the input minimum distance matrix. It calls
segment() from DNACopy with (alpha=0.001, undo.splits="sdundo", undo.SD=4)
}
\keyword{segmentMD}
